requirements.txt
crewai>=0.86.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
Flask==2.2.0
flask-cors==4.0.0
werkzeug==2.3.7
gunicorn==21.2.0
litellm>=1.55.0
pyyaml>=6.0
